pyproject.toml
[project]
name = "cartographer"
version = "0.1.0"
description = "Brownfield Cartographer — multi-agent codebase intelligence system for data-engineering repos"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "networkx>=3.0",
    "sqlglot>=20.0",
    "tree-sitter>=0.21",
    "tree-sitter-python>=0.21",
    "tree-sitter-javascript>=0.21",
    "tree-sitter-yaml>=0.0.9",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"
